InDev 0.1 MyCraft
This is the InDev version 0.1 of my "MyCraft" Mod which will be based of
basic Minecraft and Forge code that any coder will need to start off.
This is my first "Commit" I don't know how to use GitHub and this is my
first time please take it easy on me. Thank you.